refactor: :bulb:  move task classes to tasks directory
