+++ flow/Dockerfile
@@ -4,6 +4,7 @@ LABEL maintainer="shin@sforzando.co.jp"
 
 ARG PREFECT_PROJECT_NAME=etude-Prefect
 
+ENV PYTHONPATH "${PYTHONPATH}:/flow/tasks/"
 ENV WORK_DIR /flow
 WORKDIR $WORK_DIR
 ENV PREFECT__USER_CONFIG_PATH $WORK_DIR/config.toml
