+++ flow/tasks/idetail/delete_contents_task.py
@@ -1,0 +1,5 @@
+from prefect import Task
+
+class DeleteContentsTask(Task):
+    def run(self, resource_data: dict):
+        self.logger.info(f"{self.__class__.__name__}: {resource_data=}")
