+++ flow/tasks/idetail/get_paths_of_master_csv_task.py
@@ -1,0 +1,6 @@
+from prefect import Task
+
+class GetPathsOfMasterCsvTask(Task):
+    def run(self):
+        self.logger.info(f"{self.__class__.__name__}")
+        return ["csv/idetail_master.csv"]
