+++ flow/tasks/idetail/get_products_task.py
@@ -1,0 +1,6 @@
+from prefect import Task
+
+class GetProductsTask(Task):
+    def run(self):
+        self.logger.info(f"{self.__class__.__name__}")
+        return ['DUM', 'BEL']
