+++ flow/tasks/idetail/update_resources_by_product_task.py
@@ -1,0 +1,8 @@
+from typing import List
+
+from prefect import Task
+
+class UpdateResourcesByProductTask(Task):
+    def run(self, product_name: str, resource_data: List[dict]):
+        self.logger.info(f"{self.__class__.__name__}: {product_name=}, {resource_data=}")
+
