+++ flow/tasks/idetail/update_status_by_s3_raw_data_path_task.py
@@ -1,0 +1,7 @@
+from typing import List
+
+from prefect import Task
+
+class UpdateStatusByS3RawDataPathTask(Task):
+    def run(self, resource_data: List[dict]):
+        self.logger.info(f"{self.__class__.__name__}: {resource_data=}")
